chore(fractal): move src files to fain-src package
Fractal requires the preview layouts to be a component in the component directory. I wanted the preview layout to dynamically load all the files from the public folder, so that they wouldn't have to manually be added. The simplest solution was to move the src files to the fain-src package so that all the context of the public directory could live in the same package.

Other options could be:
* pass in the list of src files as the context of the preview layout.
* dynamically update the static references to the stylesheet files in the public folder